requirements.txt
aiofiles==24.1.0
aiosqlite==0.21.0
annotated-types==0.7.0
anyio==4.8.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
databases==0.9.0
distlib==0.3.9
fastapi==0.115.8
filelock==3.17.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
Jinja2==3.1.5
MarkupSafe==3.0.2
packaging==24.2
pipenv==2024.4.1
platformdirs==4.3.6
pydantic==2.10.6
pydantic_core==2.27.2
python-dotenv==1.0.1
python-multipart==0.0.20
requests==2.32.3
setuptools==75.8.0
sniffio==1.3.1
SQLAlchemy==2.0.38
starlette==0.45.3
templates==0.0.5
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
virtualenv==20.29.1
